Refactor: rename entity to domain
