+++ src/main/java/com/palindrome/studit/domain/study/domain/StudyMission.java
@@ -1,4 +1,4 @@
-package com.palindrome.studit.domain.study.entity;
+package com.palindrome.studit.domain.study.domain;
 
 public enum StudyMission {
     VELOG,
