+++ src/main/java/com/palindrome/studit/domain/user/application/AuthService.java
@@ -1,9 +1,9 @@
 package com.palindrome.studit.domain.user.application;
 
-import com.palindrome.studit.domain.user.entity.OAuthInfo;
-import com.palindrome.studit.domain.user.entity.OAuthProviderType;
-import com.palindrome.studit.domain.user.entity.User;
-import com.palindrome.studit.domain.user.entity.UserRoleType;
+import com.palindrome.studit.domain.user.domain.OAuthInfo;
+import com.palindrome.studit.domain.user.domain.OAuthProviderType;
+import com.palindrome.studit.domain.user.domain.User;
+import com.palindrome.studit.domain.user.domain.UserRoleType;
 import com.palindrome.studit.domain.user.dao.OAuthInfoRepository;
 import com.palindrome.studit.domain.user.dao.UserRepository;
 import jakarta.transaction.Transactional;
