+++ src/main/java/com/palindrome/studit/domain/user/dao/OAuthInfoRepository.java
@@ -1,6 +1,6 @@
 package com.palindrome.studit.domain.user.dao;
 
-import com.palindrome.studit.domain.user.entity.OAuthInfo;
+import com.palindrome.studit.domain.user.domain.OAuthInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
