+++ src/main/java/com/palindrome/studit/domain/user/dao/UserRepository.java
@@ -1,6 +1,6 @@
 package com.palindrome.studit.domain.user.dao;
 
-import com.palindrome.studit.domain.user.entity.User;
+import com.palindrome.studit.domain.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
