+++ src/main/java/com/palindrome/studit/domain/user/domain/UserRoleType.java
@@ -1,0 +1,5 @@
+package com.palindrome.studit.domain.user.domain;
+
+public enum UserRoleType {
+    USER, ADMIN;
+}
