--- src/main/java/com/palindrome/studit/domain/user/entity/UserRoleType.java
@@ -1,5 +1,0 @@
-package com.palindrome.studit.domain.user.entity;
-
-public enum UserRoleType {
-    USER, ADMIN;
-}
