+++ src/main/java/com/palindrome/studit/global/config/security/dto/CustomOAuth2User.java
@@ -1,6 +1,6 @@
 package com.palindrome.studit.global.config.security.dto;
 
-import com.palindrome.studit.domain.user.entity.User;
+import com.palindrome.studit.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
